LayoutTests/fast/js/kde/parse-expected.txt

PASS res is 11
PASS caught is true
PASS successfullyParsed is true
